Warning fix for wxDateTime_t <-> MSW data exchange.
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@29847 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/common/filename.cpp
@@ -237,8 +237,8 @@ static void ConvertWxToFileTime(FILETIME *ft, const wxDateTime& dt)
 {
     SYSTEMTIME st;
     st.wDay = dt.GetDay();
-    st.wMonth = dt.GetMonth() + 1;
-    st.wYear = dt.GetYear();
+    st.wMonth = (WORD)(dt.GetMonth() + 1);
+    st.wYear = (WORD)dt.GetYear();
     st.wHour = dt.GetHour();
     st.wMinute = dt.GetMinute();
     st.wSecond = dt.GetSecond();
